Extract error classification to src/download_errors + add tests
- Drop the dead "read timed out" marker (already covered by "timed out").
- Move the pure error-classification logic (DownloadErrorKind, marker
  table, classify_download_error, is_retriable_error, ERROR_USER_MESSAGES)
  out of src/bot.py into src/download_errors.py — no yt-dlp/aiogram deps,
  so it imports and unit-tests in isolation. bot.py imports from it.
- Add tests/test_download_errors.py: classification + retriability for the
  real rehydration error and each kind, case-insensitivity, ordering
  (specific kind wins over a later generic substring), and a completeness
  check that every non-UNKNOWN kind has markers and a user message. Runs
  standalone (python tests/...) or under pytest.
