+++ src/bot.py
@@ -4,10 +4,14 @@ import contextlib
 import glob
 import json
 import re
-from enum import Enum
 from datetime import datetime
 from pathlib import Path
 import yt_dlp
+from src.download_errors import (
+    classify_download_error,
+    is_retriable_error,
+    ERROR_USER_MESSAGES,
+)
 import os
 import asyncio
 import time
@@ -22,138 +26,6 @@ LIMIT_SIZE_UPL_VIDEO = 49
 # succeeds on the next attempt). Backoff grows: RETRY_BASE_DELAY * attempt.
 MAX_DOWNLOAD_ATTEMPTS = 4
 RETRY_BASE_DELAY = 1.5
-
-class DownloadErrorKind(Enum):
-    """Category of a yt-dlp download failure, inferred from its error text.
-
-    Drives three things: whether we retry (transient kinds only), what message
-    the user sees, and how the failure is tagged in logs/download_issues.jsonl.
-    """
-
-    AGE_RESTRICTED = "age_restricted"  # site wants sign-in to confirm age
-    LOGIN_REQUIRED = "login_required"  # private / sign-in / bot-check
-    UNAVAILABLE = "unavailable"  # removed / deleted / private / not found
-    GEO_BLOCKED = "geo_blocked"  # not available in this country
-    RATE_LIMITED = "rate_limited"  # HTTP 429 / too many requests
-    UNSUPPORTED = "unsupported"  # unsupported URL / no extractor / no formats
-    NETWORK = "network"  # timeout / connection / 5xx (transient)
-    EXTRACTION = "extraction"  # rehydration / "unable to extract" (flaky)
-    UNKNOWN = "unknown"  # anything we could not classify
-
-
-# Ordered most-specific → most-generic: the first kind with a matching
-# substring wins, so AGE/LOGIN/UNAVAILABLE are checked before the broad
-# EXTRACTION markers. Match is a case-insensitive substring of the message.
-_ERROR_KIND_MARKERS: dict[DownloadErrorKind, tuple[str, ...]] = {
-    DownloadErrorKind.AGE_RESTRICTED: (
-        "confirm your age",
-        "age-restricted",
-        "age restricted",
-        "inappropriate for some users",
-    ),
-    DownloadErrorKind.LOGIN_REQUIRED: (
-        "sign in to confirm you're not a bot",
-        "sign in",
-        "log in to",
-        "login required",
-        "requires authentication",
-        "private video",
-        "this video is private",
-        "this account is private",
-        "use --cookies",
-    ),
-    DownloadErrorKind.GEO_BLOCKED: (
-        "not available in your country",
-        "in your country",
-        "geo restricted",
-        "geo-restricted",
-    ),
-    DownloadErrorKind.RATE_LIMITED: (
-        "http error 429",
-        "too many requests",
-        "rate-limit",
-        "rate limit",
-    ),
-    DownloadErrorKind.UNAVAILABLE: (
-        "video unavailable",
-        "no longer available",
-        "has been removed",
-        "was deleted",
-        "this post may not be available",
-        "http error 404",
-        "account has been terminated",
-    ),
-    DownloadErrorKind.UNSUPPORTED: (
-        "unsupported url",
-        "no suitable extractor",
-        "no video formats found",
-        "is not a valid url",
-    ),
-    DownloadErrorKind.NETWORK: (
-        "timed out",
-        "timeout",
-        "read timed out",
-        "connection",
-        "temporarily",
-        "http error 5",
-        "remote end closed",
-    ),
-    DownloadErrorKind.EXTRACTION: (
-        "rehydration",
-        "universal data",
-        "unable to extract",
-        "unable to download webpage",
-        "challenge",
-    ),
-}
-
-# Kinds where retrying has a real chance of succeeding.
-_RETRIABLE_KINDS = frozenset(
-    {
-        DownloadErrorKind.NETWORK,
-        DownloadErrorKind.EXTRACTION,
-        DownloadErrorKind.RATE_LIMITED,
-    }
-)
-
-# User-facing (Ukrainian) message per kind. UNKNOWN falls back to the raw error.
-_ERROR_USER_MESSAGES = {
-    DownloadErrorKind.AGE_RESTRICTED: (
-        "🔞 Відео з віковим обмеженням — сайт вимагає вхід/підтвердження віку, "
-        "тож завантажити не вдалося."
-    ),
-    DownloadErrorKind.LOGIN_REQUIRED: (
-        "🔒 Відео приватне або потребує входу — завантажити не можу."
-    ),
-    DownloadErrorKind.UNAVAILABLE: (
-        "🚫 Відео недоступне (видалене, приватне або не існує)."
-    ),
-    DownloadErrorKind.GEO_BLOCKED: "🌍 Відео недоступне у цьому регіоні.",
-    DownloadErrorKind.RATE_LIMITED: (
-        "⏳ Забагато запитів до сайту. Спробуй, будь ласка, трохи згодом."
-    ),
-    DownloadErrorKind.UNSUPPORTED: "❓ Це посилання не підтримується.",
-    DownloadErrorKind.NETWORK: (
-        "📡 Проблема з мережею під час завантаження. Спробуй ще раз."
-    ),
-    DownloadErrorKind.EXTRACTION: (
-        "⚠️ Сайт тимчасово не віддає відео. Спробуй ще раз за хвилину."
-    ),
-}
-
-
-def classify_download_error(msg: str) -> DownloadErrorKind:
-    """Map a yt-dlp error message to a DownloadErrorKind (first match wins)."""
-    low = msg.lower()
-    for kind, markers in _ERROR_KIND_MARKERS.items():
-        if any(marker in low for marker in markers):
-            return kind
-    return DownloadErrorKind.UNKNOWN
-
-
-def _is_retriable_error(msg: str) -> bool:
-    """True if the error is a transient kind worth retrying."""
-    return classify_download_error(msg) in _RETRIABLE_KINDS
 
 
 # Network/extractor robustness defaults shared by every yt-dlp call.
@@ -483,7 +355,7 @@ class Bot_Func:
                         with yt_dlp.YoutubeDL(opts) as ydl:
                             return ydl.extract_info(med_url, download=True)
                     except yt_dlp.utils.DownloadError as e:
-                        if attempt >= MAX_DOWNLOAD_ATTEMPTS or not _is_retriable_error(
+                        if attempt >= MAX_DOWNLOAD_ATTEMPTS or not is_retriable_error(
                             str(e)
                         ):
                             raise
@@ -558,7 +430,7 @@ class Bot_Func:
             )
             if strt_dwn_msg:
                 await strt_dwn_msg.delete()
-            await user_msg.reply(_ERROR_USER_MESSAGES.get(kind, f"Download error: {e}"))
+            await user_msg.reply(ERROR_USER_MESSAGES.get(kind, f"Download error: {e}"))
             return
         except Exception as e:
             self.log.error(f"❌ An error occurred: {e}")
